refactor: Create a Component for Header
