+++ src/App.tsx
@@ -5,12 +5,14 @@ import Home from "./pages/Home";
 import Login from "./pages/Login";
 import { AuthGoogleProvider } from "./contexts/authGoogle";
 import Footer from "./components/Footer";
+import Header from "./components/Header";
 
 function App() {
   return (
     <>
       <Router>
         <AuthGoogleProvider>
+          <Header />
           <Routes>
             <Route Component={Home} path='/' />
             <Route Component={RaceRegister} path='/cadastro-corrida' />
